backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "story-writers-backend"
version = "0.1.0"
description = "FastAPI backend for a multi-agent story generator"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.99.0,<0.100.0",
  "uvicorn>=0.30.0",
  "sqlmodel>=0.0.8,<0.0.9",
  "PyJWT>=2.9.0",
  "email-validator>=2.2.0",
  "httpx>=0.27.0,<0.28.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
